requirements.txt
admin==0.0.1
alembic==1.4.2
attrs==19.3.0
Babel==2.8.0
blinker==1.4
branca==0.4.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
cryptography==2.9.2
Django==1.7
dnspython==1.16.0
dominate==2.5.1
elasticsearch==7.7.1
email-validator==1.1.1
feather-format==0.4.1
Fiona==1.8.13.post1
Flask==1.1.2
Flask-Admin==1.5.6
Flask-Babel==1.0.0
Flask-BabelEx==0.9.4
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
folium==0.11.0
geofeather==0.3.0
geopandas==0.7.0
github3.py==1.3.0
guess-language-spirit==0.5.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jwcrypto==0.7
Mako==1.1.3
MarkupSafe==1.1.1
munch==2.5.0
numpy==1.18.5
pandas==1.0.4
passlib==1.7.2
Pillow==5.3.0
pyarrow==0.17.1
pycparser==2.20
PyJWT==1.7.1
pyproj==2.6.1.post1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
redis==3.5.3
requests==2.23.0
rq==1.4.2
Shapely==1.7.0
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.17
uritemplate==3.0.1
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
